fiks tester, slett deprekert kode
