+++ src/felleskomponenter/oppgaveliste/behandlingOppgave.less
@@ -52,10 +52,15 @@
     display: flex;
     flex-direction: column;
     justify-content: space-between;
+    padding-top: 0.5em;
 
     border: 1px solid #e9e9e9;
     margin-top: -2.2em;
     min-height: 7.5em;
+  }
+
+  .row {
+    margin-left: 0.25em;
   }
 }
