--- src/sider/forside/komponenter/mineoppgaver/index.js
@@ -1,3 +1,0 @@
-import MineOppgaver from "./mineoppgaver";
-
-export default MineOppgaver;
